+++ akka-stream-typed/src/main/scala/akka/stream/typed/package.scala
@@ -1,0 +1,9 @@
+/**
+ * Copyright (C) 2018 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package akka.stream.typed
+
+package object scaladsl {
+  type ActorMaterializer = akka.stream.ActorMaterializer
+}
