+++ akka-stream-typed/src/test/scala/docs/akka/stream/typed/ActorSourceSinkExample.scala
@@ -5,11 +5,14 @@
 package docs.akka.stream.typed
 
 import akka.NotUsed
-import akka.stream.ActorMaterializer
+import akka.actor.typed.ActorSystem
+import akka.stream.typed.scaladsl.ActorMaterializer
 
 object ActorSourceSinkExample {
 
-  implicit val mat: ActorMaterializer = ???
+  val system: ActorSystem[_] = ???
+
+  implicit val mat: ActorMaterializer = ActorMaterializer()(system)
 
   {
     // #actor-source-ref
